+++ CHANGELOG.md
@@ -5,10 +5,11 @@
 - Added methods to `GOMClientDelegate` protocol:
 
     ```objective-c
+    - (BOOL)gomClientShouldReconnect:(GOMClient *)gomClient
     - (BOOL)gomClient:(GOMClient *)gomClient shouldReRegisterObserverWithBinding:(GOMBinding *)binding;
     ```
 
-    which get called when the client reconnects to the GOM through its websocket.
+    which get called when the client's websocket fails / reconnects to the GOM.
 
 - Private methods are public now: